+++ spec/suspenders/app_generator_spec.rb
@@ -1,7 +1,0 @@
-require "spec_helper"
-
-RSpec.describe Suspenders::AppGenerator do
-  it "includes ExitOnFailure" do
-    expect(described_class.ancestors).to include(Suspenders::ExitOnFailure)
-  end
-end