+++ templates/partials/email_smtp.rb
@@ -1,2 +1,0 @@
-  config.action_mailer.delivery_method = :smtp
-  config.action_mailer.smtp_settings = SMTP_SETTINGS